Add files via upload
Question : Write a function, generateHistogram, which generates the histogram of an image. The function
should take an image data array (with pixel values in the range 0 – 255) as its only parameter and
return an array containing the histogram of the image. The histogram can be displayed using the
built in plotting function. Use this new function to generate and display histograms for the
following images.