Add files via upload
Create a spatial filter to get the horizontal edge of the image “two_cats.jpg”. Create a spatial filter
to get the vertical edge of the image (read the MATLAB documentation of fspecial). Now
transform both of these filters to the frequency domain. Also Transform the two_cats image to the
frequency domain. Apply the appropriate operations in the frequency domain. Transform the data
back into the spatial domain. Sum the horizontal and vertical edge components together.