Add files via upload
Consider the images and the problem of matching a given image template (ex. Image with
symbol ‘T’) to different regions/objects in a target image (the bigger image with many symbol).
Based on the techniques studied till now, implement a method to locate the region in the target
image which matches with the template image.