Update flipImage.m
Write a function flipImage which flips an image either vertically or horizontally. The function
should take two parameters – the matrix storing the image data and a flag to indicate
whether the image should be flipped vertically or horizontally. Use this function from the
command line to flip a given image both vertically and horizontally which should give the
following results.