Add files via upload
Write a program that calculates the average intensity value of the pixels in the image
Moon.bmp and then thresholds this image based on this intensity. Thresholding means that
a new image is generated in which each pixel has intensity 1.0 if the corresponding pixel in
the original image has a value above the threshold and 0 otherwise. Use this new function
from the command line on the image moon.tif .